+++ tests/testthat/test-plasmode.R
@@ -15,7 +15,8 @@ pars <- list(A = list(beta=c(-0.5,0.5,0.25,-0.5)),
              cop = list(Y = list(Z1 = list(beta=c(0.5,0.5)),
                                  Z2 = list(beta=c(0.75,0.2)))))
 
-dat <- suppressMessages(rfrugalParam(formulas = forms, family = fams, pars = pars, dat = df))
+dat <- rfrugalParam(formulas = forms, family = fams, pars = pars, dat = df,
+                    control = list(quiet=TRUE))
 
 test_that("plasmode simulation works", {
   expect_equal(dim(dat), c(n, 4L))
@@ -49,4 +50,5 @@ pars <- list(A = list(beta=c(0,1)),
              cop = list(beta=-1))
 
 dat <- data.frame(U = rep(1,100))
-dat <- suppressMessages(rfrugalParam(formulas=forms, family=fams, pars=pars, dat=dat))
+dat <- rfrugalParam(formulas=forms, family=fams, pars=pars, dat=dat,
+                    control=list(quiet=TRUE))